[IMP] product_materials module: Changed two fields from Integer to Float

See merge request tawasta/odoo/product!296

+++ product_materials/models/product_material_composition.py
@@ -49,11 +49,11 @@ class ProductMaterialComposition(models.Model):
         string="Material Sublevel",
     )
 
-    biogenic_material_weight_percentage = fields.Integer(
+    biogenic_material_weight_percentage = fields.Float(
         related="product_material_id.biogenic_material_weight_percentage"
     )
 
-    renewable_weight_percentage = fields.Integer(
+    renewable_weight_percentage = fields.Float(
         related="product_material_id.renewable_weight_percentage"
     )
